fix(prisma): :recycle: Move prisma files due to fix breaking changes in v6.6.0
