+++ package.json
@@ -82,6 +82,6 @@
     "typescript-eslint": "^8.26.1"
   },
   "prisma": {
-    "schema": "./schema"
+    "schema": "./prisma/schema"
   }
 }
