+++ src/config/prisma/prisma.service.ts
@@ -3,7 +3,7 @@ import {
   type OnModuleDestroy,
   type OnModuleInit,
 } from '@nestjs/common';
-import { PrismaClient } from '@prisma/client';
+import { PrismaClient } from '@prisma';
 
 @Injectable()
 export class PrismaService
